requirements.txt
<<<<<<< HEAD
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Agent Framework
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langgraph==0.0.20

# NLP & ML
transformers==4.36.0
sentence-transformers==2.2.2
spacy==3.7.2
torch==2.1.1

# Document Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
pillow==10.1.0

# Vector Database
qdrant-client==1.7.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Task Queue
celery==5.3.4

# HTTP & API
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
pandas==2.1.4

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Add pandas to requirements
pandas==2.1.4
python-dateutil==2.8.2

# Add fpdf library for PDF generation
fpdf==1.7.2

=======
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coverage==7.10.0
dataclasses-json==0.6.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
filetype==1.2.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mypy_extensions==1.1.0
numpy==2.3.1
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.2.1
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.2.0
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ulid==1.1
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.23.0
>>>>>>> fastmvc/main
